First attempt of skipping unwanted functions #5
This commits adds the source code for the PoC binary that is going to be fuzzed and its Python script to run it with Qiling. In the script, the binary is run normally until reaching the function desired to be fuzzed, when this happens the function's parameter is changed for something else. This will be helpful in the future to skip the "sleep" call and fuzz the rest with an arbitrary fuzzed string as input.