Add harness and corpus for cJSON #2
This commit adds a fuzzing harness and corpus to allow fuzzing of the cJSON .so file located in Wyze Cam V3's firmware (version 4.36.8.26, cJSON version 1.7.13). This library was compiled for the mipsel architecture thus this fuzzing setup being created for fuzzing cross-architecture binaries with AFL++ QEMU mode.

The JSON files that make up the corpus are the ones provided in cJSON's official [repository](https://github.com/DaveGamble/cJSON).